Refactor: gfi_new -> gfi
