+++ gen-rs/src/inference/mod.rs
@@ -1,9 +1,10 @@
+// pub mod old_importance;
 pub mod importance;
-pub mod new_importance;
-pub mod mh;
-pub mod particle_filter;
+// pub mod mh;
+// pub mod particle_filter;
 
+// pub use old_importance::importance_sampling;
 pub use importance::importance_sampling;
-pub use mh::metropolis_hastings;
-pub use mh::mh;
-pub use particle_filter::ParticleSystem;
+// pub use mh::metropolis_hastings;
+// pub use mh::mh;
+// pub use particle_filter::ParticleSystem;
