+++ gen-rs/tests/pointed/mod.rs
@@ -1,8 +1,8 @@
-mod trace;
-mod model;
-mod proposal;
+// mod trace;
+// mod model;
+// mod proposal;
 pub mod types_2d;
 
-pub use trace::PointedTrace;
-pub use model::PointedModel;
-pub use proposal::DriftProposal;
+// pub use trace::PointedTrace;
+// pub use model::PointedModel;
+// pub use proposal::DriftProposal;
